+++ src/lobby.rs
@@ -1,8 +1,8 @@
 use crate::{
     error::CardsError,
-    types::{ChargingRules, Event, GameId, Participant, Player, UserId},
+    lobby::event::LobbyEvent,
+    types::{ChargingRules, GameId, Participant, Player, UserId},
 };
-use serde::{Deserialize, Serialize};
 use std::{
     collections::{hash_map::Entry, HashMap, HashSet, VecDeque},
     sync::Arc,
@@ -12,6 +12,9 @@ use tokio::sync::{
     Mutex,
 };
 
+pub mod endpoints;
+pub mod event;
+
 #[derive(Clone)]
 pub struct Lobby {
     inner: Arc<Mutex<Inner>>,
@@ -20,50 +23,6 @@ pub struct Lobby {
 struct Inner {
     subscribers: HashMap<UserId, Vec<UnboundedSender<LobbyEvent>>>,
     games: HashMap<GameId, HashSet<Participant>>,
-}
-
-#[serde(tag = "type", rename_all = "snake_case")]
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
-pub enum LobbyEvent {
-    Ping,
-    JoinLobby {
-        user_id: UserId,
-    },
-    NewGame {
-        game_id: GameId,
-        user_id: UserId,
-    },
-    LobbyState {
-        subscribers: HashSet<UserId>,
-        games: HashMap<GameId, Vec<Player>>,
-    },
-    JoinGame {
-        game_id: GameId,
-        player: Player,
-    },
-    LeaveGame {
-        game_id: GameId,
-        user_id: UserId,
-    },
-    FinishGame {
-        game_id: GameId,
-    },
-    Chat {
-        user_id: UserId,
-        message: String,
-    },
-    LeaveLobby {
-        user_id: UserId,
-    },
-}
-
-impl Event for LobbyEvent {
-    fn is_ping(&self) -> bool {
-        match self {
-            LobbyEvent::Ping => true,
-            _ => false,
-        }
-    }
 }
 
 impl Lobby {
