+++ src/macros.rs
@@ -1,9 +1,9 @@
 macro_rules! infallible {
-    ($t:ty) => {impl warp::Filter<Extract = ($t,), Error = std::convert::Infallible> + Clone + Send}
+    ($t:ty) => {impl warp::Filter<Extract = ($t,), Error = std::convert::Infallible> + Clone + Send + Sync + 'static}
 }
 
 macro_rules! rejection {
-    ($t:ty) => {impl warp::Filter<Extract = ($t,), Error = warp::Rejection> + Clone + Send}
+    ($t:ty) => {impl warp::Filter<Extract = ($t,), Error = warp::Rejection> + Clone + Send + Sync + 'static}
 }
 
 macro_rules! reply {
